Catch situation where intersection is false
When sweeping, can jump from former closest to new closest and the two may not connect nor one in front of the other. In that case, there are likely two polygons that do not overlap or obscure one another, and we have jumped to the other. Intersection is then false.