Working tests of rendering terrain elevation texture
Important to use the correct variable name for the texture!

No real difference between uploading the entire texture versus cutting out the frame. Whole texture is easier to deal with. Might be more of a performance hit for a larger texture. Current EV tex is 0.25 resolution and only the red channel.